// js/app/src/pages/project/__generated__/ProjectStats_project.graphql.ts
export type ProjectStats_project$data = {
  readonly costSummary: {
    readonly completion: {
      readonly cost: number | null;
    };
    readonly prompt: {
      readonly cost: number | null;
    };
    readonly total: {
      readonly cost: number | null;
    };
  };
  readonly documentEvaluationNames: ReadonlyArray<string>;
  readonly id: string;
  readonly latencyMsP50: number | null;
  readonly latencyMsP99: number | null;
  readonly spanAnnotationNames: ReadonlyArray<string>;
  readonly timeRangeTraceCount: number;
  readonly " $fragmentType": "ProjectStats_project";
};
export type ProjectStats_project$key = {
  readonly " $data"?: ProjectStats_project$data;
  readonly " $fragmentSpreads": FragmentRefs<"ProjectStats_project">;
};